requirements.txt
django
djangorestframework 
django-cors-headers
pillow
django-admin-sortable2
djangorestframework-simplejwt